refactor(pointer): replace `schema` with `schemas`
