+++ libraries/pointer/package.json
@@ -25,7 +25,7 @@
   "dependencies": {
     "@toa.io/console": "0.6.0",
     "@toa.io/generic": "0.7.0",
-    "@toa.io/schema": "0.7.2-dev.0",
+    "@toa.io/schemas": "0.7.2-dev.0",
     "@toa.io/yaml": "0.7.2-dev.0"
   }
 }
