--- libraries/pointer/source/uris/.construct/index.js
@@ -1,7 +1,0 @@
-'use strict'
-
-const { normalize } = require('./normalize')
-const { validate } = require('./validate')
-
-exports.normalize = normalize
-exports.validate = validate
