+++ libraries/pointer/source/uris/index.js
@@ -2,6 +2,8 @@
 
 const { construct } = require('./construct')
 const { resolve } = require('./resolve')
+const { validate } = require('./validate')
 
 exports.construct = construct
 exports.resolve = resolve
+exports.validate = validate
